feat: include packages from custom local repository
This is needed to install an older version of lvm2, which is no longer
on Void's servers, so I needed to recompile it with
void-packages/xbps-src.

See this issue[0] and a workaround[1], although that workaround only
works for Arch, since Void doesn't use lvmetad (or at least I wasn't
able to get it working after starting the service outside of chroot).

[0]: https://bugs.archlinux.org/task/61040
[1]: https://bbs.archlinux.org/viewtopic.php?pid=1820949#p1820949