feat: abandon GPG for LUKS key encryption
GPG was more trouble than what a slight security advantage (decrypting
the LUKS key with a GPG smartcard) was worth it. In practice, I've
once managed to lock myself out of a system by mistakenly entering
my GPG passphrase 3 times, which factory resets the card  -__- (a good
security measure TBF, just annoying for my use case). If I didn't have
a local backup card, I would've been SoL, though I ended up hosing the
system afterwards anyway.

In any case, it's too brittle of a system to work in practice,
especially when NVIDIA drivers disable the drm initramfs module, which
is apparently needed for the GnuPG pinentry program... So, this avoids
that whole mess, and sticks with a plain LUKS passphrase to encrypt the
data partition. /boot is still unencrypted, until I find a sensible way
to configure GRUB to decrypt the volume directly, or implement
passphrase caching so that I don't need to enter the passhprase more
than once, which is a problem for the system this fork is based on.

Or switch to fucking BSD.